/login/ is working, added secured routes and added tests

+++ routes/index.js
@@ -5,56 +5,27 @@
 const express = require('express');
 const router = express.Router();
 const alasql = require('alasql');
-//require('../item_db.js');
-const db = require('../item_db.js');
+require('../item_db.js');
+require('../config/auth.js');
 const passport = require('passport');
-const JsonStrategy = require('passport-json').Strategy;
 const bodyParser = require('body-parser');
-const flash = require('connect-flash');
-
+const session = require('express-session');
 // ----- configuration -----
 router.use(require('morgan')('combined'));
-router.use(require('cookie-parser')());
 router.use(bodyParser.json());
-router.use(require('express-session')({ secret: 'keyboard cat', resave: false, saveUninitialized: false }));
+router.use(session({ secret: 'keyboard cat', resave: false, saveUninitialized: true }));
 router.use(passport.initialize());
 router.use(passport.session());
-router.use(flash());
 
-passport.serializeUser(function(user, cb) {
-	cb(null, user.id);
-});
-
-passport.deserializeUser(function(id, done) {
-	db.findById(id, function(err, user) {
-		done(err, user);
-	});
-});
-
-passport.use(new JsonStrategy(
-	function(username, password, cb) {
-		db.findByUsername(username, function(err, user) {
-			if (err) {
-				return cb(err);
-			}
-			if (!user){
-				return cb(null, false);
-			}
-			if (user.password != password) {
-				return cb(true, null);
-			}
-			return cb(null, user);
-		});
-	}));
-
+// ----- funcion for checking  -----
 function isLoggedIn(req, res, next) {
 	if(req.isAuthenticated()) {
 		return next();
 	}
 	return res.status(403).send('You are not logged in');
-	//return res.redirect('/login');
 }
 
+// ----- middleware -----
 router.param('itemId', (req, res, next, id) => {
 	const itemId = Number(id);
 	if ( isNaN(itemId)){
@@ -71,28 +42,31 @@ router.param('itemId', (req, res, next, id) => {
 	}
 });
 
-// Get all Items
-router.get('/items/', (req, res, next) => {
-	const items = alasql('SELECT * FROM items');
+//===============================================
+//====           Item routes                 ====
+//===============================================
+
+// ----- Get all Items -----
+router.get('/items/', isLoggedIn, (req, res, next) => {
+	const items = alasql('SELECT * FROM items WHERE userId=?', req.user.id);
 	res.send(items);
 	next();
 });
 
-// Get all Users
-router.get('/users/', (req, res, next) => {
-	const users = alasql('SELECT * FROM users');
-	res.send(users);
-	next();
-});
-
 // Get a single Item
-router.get('/items/:itemId', (req, res, next) => {
-	const item = alasql('SELECT * FROM items WHERE itemId= ?', req.itemId);
-	res.send(item);
+router.get('/items/:itemId', isLoggedIn, (req, res, next) => {
+	const userId = alasql('SELECT userId FROM items WHERE itemId='+ req.itemId);
+	const id = userId[0];
+	if (id.userId === req.user.id ){
+		const item = alasql('SELECT * FROM items WHERE itemId='+ req.itemId + ' AND userId='+ req.user.id);
+		res.send(item);
+	} else{
+		res.status(403).send('That item does not belong to you');
+	}
 	next();
 });
 
-// Create a new item
+// ----- Create a new item -----
 router.post('/items/', (req, res, next) => {
 	const newItem = req.body;
 	if (isNaN(newItem.quantity) ){
@@ -103,31 +77,49 @@ router.post('/items/', (req, res, next) => {
 	next();
 });
 
-// Delete an Item
-router.delete('/items/:itemId', (req, res, next) => {
-	alasql('DELETE FROM items WHERE itemId=?', req.itemId);
+// ----- Delete an Item -----
+router.delete('/items/:itemId', isLoggedIn, (req, res, next) => {
+	const userId = alasql('SELECT userId FROM items WHERE itemId='+ req.itemId);
+	const id = userId[0];
+	if (id.userId === req.user.id ){
+		alasql('DELETE FROM items WHERE itemId='+ req.itemId + ' AND userId='+ req.user.id);
+		res.status(204).send();
+	} else{
+		res.status(403).send('That item does not belong to you');
+	}
+	next();
+});
+
+// ----- Delete all Items -----
+router.delete('/items/', isLoggedIn, (req, res, next) => {
+	alasql('DELETE FROM items WHERE userId=?', req.user.id);
 	res.status(204).send();
 	next();
 });
 
-// Delete all Items
-router.delete('/items/', (req, res, next) => {
-	alasql('DELETE FROM items');
-	res.status(204).send();
-	next();
-});
-
-router.post('/login/', function(req, res, next) {
-	passport.authenticate('json', function(err, user) {
-		if (err) {
-			return res.status(403).send('Invalid username or password');
-		}
-		return res.status(200).send('Hello user ' + user.username);
+//===============================================
+//====           User routes                 ====
+//===============================================
+// ----- Login an User and start session -----
+router.post('/login/', (req, res, next) => {
+	passport.authenticate('json', (err, user) => {
+		req.logIn(user, (err) => {
+			if (err) { return res.status(403).send('Invalid username or password'); }
+			return res.status(200).send('Hello user ' + user.username);
+		});
 	})(req, res, next);
 });
 
-router.get('/profile/', isLoggedIn, function(req, res){
-	return res.status(200).send(req.user.username);
+// ----- Get profile of the User -----
+router.get('/profile/', isLoggedIn, (req, res) =>{
+	return res.status(200).send(alasql('SELECT id, username FROM users WHERE id= ?', req.user.id));
+});
+
+// ----- Logout the User and end session -----
+router.get('/logout', isLoggedIn, (req, res) => {
+	req.logout();
+	req.session.destroy();
+	res.send('ok');
 });
 
 module.exports = router;
